Move dev command
