+++ packages/@tinacms/cli/src/next/commands/dev-command/server/index.ts
@@ -3,7 +3,7 @@ import bodyParser from 'body-parser'
 import cors from 'cors'
 import { resolve as gqlResolve, Database } from '@tinacms/graphql'
 import { tinaTailwind } from '../tailwind'
-import { ConfigManager } from '../../config-manager'
+import { ConfigManager } from '../../../config-manager'
 import { createMediaRouter } from './routes'
 import { parseMediaFolder } from './models/media'
 import { transform } from 'esbuild'
